+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Added `\Wizaplace\SDK\Vendor\Order\OrderService::listOrders` and associated classes
 - Added `\Wizaplace\SDK\Pim\Product\Attribute\AttributeService::getProductAttributes` and associated classes
 - Added `\Wizaplace\SDK\Pim\Product\Attribute\AttributeService::getProductAttribute` and associated classes
+- Added `\Wizaplace\SDK\Pim\Product\Attribute\AttributeService::getCategoryAttributes` and associated classes
 - Added `\Wizaplace\SDK\Pim\Option\OptionService::getCategoryOptions` and associated classes
 
 ### Bugfixes